+++ CHANGELOG.md
@@ -9,8 +9,7 @@
 * Upgrade to Metabase 0.33.2
 * New Crossdata driver as a plugin
 * [ROCK-629] Add JWT authentication
-
-
+* Fix: Revert too strict site-url validation
 
 ## 0.34.0-7f4b16c (Built: September 19, 2019 | Released: September 20, 2019)